requirements.txt
# Core 
python-dotenv==1.0.0
langchain>=0.1.10,<0.2.0
langchain_groq==0.2.4

# Web scraping and navigation
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1

# PDF handling
PyPDF2==3.0.1
pdf2image==1.16.3
PyMuPDF==1.23.7
pytesseract==0.3.10

# Data processing
pandas==2.1.3
numpy==1.26.2

# UI
streamlit==1.29.0

# Agentic tools
crewai==0.28.8
# Ensure compatible versions with crewai
langchain-core>=0.1.10
langchain-community>=0.0.16